requirements.txt
python-docx
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
langchain-community
requests
python-dotenv
duckduckgo-search
fuzzywuzzy
pyautogui
pynput
python-Levenshtein
pywin32
livekit
mem0ai

playwright
pydub
beautifulsoup4